Tool to customize anspress capabilities
